agentic_rag/requirements.txt
pandas
langchain
langchain-community
sentence-transformers
faiss-cpu
smolagents 
datasets